requirements.txt
Flask==3.0.3
pandas==2.2.2
scikit-learn==01.3.2
fuzzywuzzy==0.18.0
numpy==1.26.2
google-generativeai==0.4.0